Reworked TessierStyles API
In the old situation, styles would be specified as strings, possibly
with parameters that were also specified in a string format. A lot of
bookkeeping was involved to apply all this correctly. I have reworked
the styles API into a cleaner one where every style is an object doing
its own bookkeeping so no external bookkeeping is needed anymore. This
should make the implementation of new well-behaving styles easier,
especially if they have parameters.